requirements.txt
python==3.11
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
transformers==4.43.3
datasets==2.20.0
accelerate==0.31.0
evaluate==0.4.2
bert-score==0.3.13
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
tqdm==4.66.4
matplotlib==3.9.0
seaborn==0.13.2
pyyaml==6.0.1
rich==13.7.1
typing-extensions==4.12.2